+++ src/error/type.js
@@ -142,6 +142,10 @@ export default {
     code: 10028,
     message: 'Database error',
   },
+  WAITING_FOR_DPKI_CALLBACK_URL_SET: {
+    code: 10029,
+    message: 'Waiting for DPKI callback URLs to be set',
+  },
 
   // Client errors
   PATH_PARAMS_VALIDATION_FAILED: {