+++ src/routes/middleware/ready_handler.js
@@ -22,10 +22,13 @@
 
 import * as tendermint from '../../tendermint';
 import { registeredMsqAddress } from '../../core/common';
+import { isCallbackUrlsSet } from '../../utils/external_crypto_service';
 
 import errorType from '../../error/type';
 
 import logger from '../../logger';
+
+import * as config from '../../config';
 
 export default function readyHandler(req, res, next) {
   // Reject all requests when tendermint is not yet ready.
@@ -61,12 +64,28 @@ export default function readyHandler(req, res, next) {
     return;
   }
 
-  // Reject all POST calls while message queue address is being registered
   if (
-    !registeredMsqAddress() &&
+    config.useExternalCryptoService &&
+    !isCallbackUrlsSet() &&
     req.method === 'POST' &&
     !req.url.endsWith('/dpki/node/callback')
   ) {
+    const responseBody = {
+      error: {
+        errorType: errorType.WAITING_FOR_DPKI_CALLBACK_URL_SET.message,
+        code: errorType.WAITING_FOR_DPKI_CALLBACK_URL_SET.code,
+      },
+    };
+    res.status(503).json(responseBody);
+    logger.error({
+      message: 'Responded Service Unavailable with HTTP code 503',
+      responseBody,
+    });
+    return;
+  }
+
+  // Reject all POST calls while message queue address is being registered
+  if (!registeredMsqAddress() && req.method === 'POST') {
     const responseBody = {
       error: {
         message: errorType.REGISTERING_MESSAGE_QUEUE_ADDRESS.message,
@@ -81,11 +100,7 @@ export default function readyHandler(req, res, next) {
     return;
   }
 
-  if (
-    !tendermint.expectedTxsLoaded &&
-    req.method === 'POST' &&
-    !req.url.endsWith('/dpki/node/callback')
-  ) {
+  if (!tendermint.expectedTxsLoaded && req.method === 'POST') {
     const responseBody = {
       error: {
         message: errorType.LOADING_EXPECTED_TXS_CACHE.message,